+++ src/Dialog.jsx
@@ -135,6 +135,8 @@ Dialog.propTypes = {
   onOk: PropTypes.func,
   htmlClassName: PropTypes.string,
   getContainer: PropTypes.func,
+  className: PropTypes.string,
+  iconClassName: PropTypes.string,
 };
 
 Dialog.defaultProps = {
@@ -155,9 +157,14 @@ Dialog.defaultProps = {
   htmlClassName: '',
 };
 
+function adjustIcon(props, defaultIcon) {
+  const icon = props.iconClassName ? <Icon name={props.iconClassName} /> : <Icon name={defaultIcon} />;
+  return props.icon ? props.icon : icon;
+}
+
 Dialog.info = (props) => {
   assign(props, {
-    icon: <Icon name="tishi-full" />,
+    icon: adjustIcon(props, 'tishi-full'),
     okCancel: false,
   });
   return confirm(props, Dialog);
@@ -165,7 +172,7 @@ Dialog.info = (props) => {
 
 Dialog.success = (props) => {
   assign(props, {
-    icon: <Icon name="chenggong-full" />,
+    icon: adjustIcon(props, 'chenggong-full'),
     okCancel: false,
   });
   return confirm(props, Dialog);
@@ -173,7 +180,7 @@ Dialog.success = (props) => {
 
 Dialog.error = (props) => {
   assign(props, {
-    icon: <Icon name="biaodanlei-tongyongqingchu" />,
+    icon: adjustIcon(props, 'biaodanlei-tongyongqingchu'),
     okCancel: false,
   });
   return confirm(props, Dialog);
@@ -181,7 +188,7 @@ Dialog.error = (props) => {
 
 Dialog.confirm = (props) => {
   assign(props, {
-    icon: <Icon name="jinggao-full" />,
+    icon: adjustIcon(props, 'jinggao-full'),
     okCancel: true,
   });
   return confirm(props, Dialog);